pyproject.toml
[project]
name = "django-react-starter"
version = "0.1.0"
description = "Starter Template for Django and React"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "python-dotenv>=1.1.0",
    "whitenoise>=6.9.0",
]
